DPTS-558 Refactor part one - all tests pass coverage unchanged
