+++ application/hk/symptomdiscriminators/handler.py
@@ -1,7 +1,6 @@
-import csv
 import psycopg2
 import psycopg2.extras
-from utilities import logger, message, common
+from utilities import logger, message, common, database
 from datetime import datetime
 
 csv_column_count = 3
@@ -20,9 +19,9 @@ def request(event, context):
     filename = event["filename"]
     bucket = event["bucket"]
     initialise_summary_count()
-    db_connection = common.connect_to_database(env, event, start)
+    db_connection = database.connect_to_database(env, event, start)
     csv_file = common.retrieve_file_from_bucket(bucket, filename, event, start)
-    lines = process_file(csv_file, event, start)
+    lines = common.process_file(csv_file, event, start, 3)
     for row, values in lines.items():
         if check_table_for_id(db_connection, row, values, filename, event, start):
             query, data = generate_db_query(values, event, start)
@@ -36,19 +35,19 @@ def request(event, context):
     return "Symptom discriminators execution successful"
 
 
-def process_file(csv_file, event, start):
-    lines = {}
-    count = 0
-    csv_reader = csv.reader(csv_file.split("\n"))
-    for line in csv_reader:
-        count += 1
-        if not common.check_csv_format(line, csv_column_count):
-            logger.log_for_error("Incorrect line format, should be 3 but is {}".format(len(line)))
-        else:
-            lines[str(count)] = {"id": line[0], "description": line[1], "action": line[2]}
-    if lines == {}:
-        message.send_failure_slack_message(event, start)
-    return lines
+# def process_file(csv_file, event, start):
+#     lines = {}
+#     count = 0
+#     csv_reader = csv.reader(csv_file.split("\n"))
+#     for line in csv_reader:
+#         count += 1
+#         if not common.check_csv_format(line, csv_column_count):
+#             logger.log_for_error("Incorrect line format, should be 3 but is {}".format(len(line)))
+#         else:
+#             lines[str(count)] = {"id": line[0], "description": line[1], "action": line[2]}
+#     if lines == {}:
+#         message.send_failure_slack_message(event, start)
+#     return lines
 
 
 def generate_db_query(row_values, event, start):
