+++ application/hk/symptomgroups/handler.py
@@ -1,7 +1,6 @@
-import csv
 import psycopg2
 import psycopg2.extras
-from utilities import logger, message, common
+from utilities import logger, message, common, database
 from datetime import datetime
 
 csv_column_count = 3
@@ -19,10 +18,11 @@ def request(event, context):
     env = event["env"]
     filename = event["filename"]
     bucket = event["bucket"]
-    initialise_summary_count()
-    db_connection = common.connect_to_database(env, event, start)
+    summary_count_dict = common.initialise_summary_count()
+    db_connection = database.connect_to_database(env, event, start)
     csv_file = common.retrieve_file_from_bucket(bucket, filename, event, start)
-    extracted_data = extract_data_from_file(csv_file)
+    csv_data = common.process_file(csv_file, event, start, 3)
+    extracted_data = extract_query_data_from_csv(csv_data)
     process_extracted_data(db_connection, extracted_data)
     logger.log_for_audit(
         "Symptom groups updated: {0}, inserted: {1}, deleted: {2}".format(
@@ -34,30 +34,30 @@ def request(event, context):
 
 
 # TODO consider moving to common ultimately
-def extract_data_from_file(csv_file):
-    lines = {}
-    count = 0
-    csv_reader = csv.reader(csv_file.split("\n"))
-    for line in csv_reader:
-        count += 1
-        if len(line) > 0:
-            query_data = extract_query_data_from_csv(line)
-            if len(query_data) != data_column_count:
-                logger.log_for_audit(
-                    "Problem constructing data from line {} of csv expecting {} items but have {}".format(
-                        str(count), str(csv_column_count), str(len(line))
-                    ),
-                )
-            else:
-                lines[str(count)] = query_data
-    return lines
+# def process_file(csv_file):
+#     lines = {}
+#     count = 0
+#     csv_reader = csv.reader(csv_file.split("\n"))
+#     for line in csv_reader:
+#         count += 1
+#         if len(line) > 0:
+#             query_data = extract_query_data_from_csv(line)
+#             if len(query_data) != data_column_count:
+#                 logger.log_for_audit(
+#                     "Problem constructing data from line {} of csv expecting {} items but have {}".format(
+#                         str(count), str(csv_column_count), str(len(line))
+#                     ),
+#                 )
+#             else:
+#                 lines[str(count)] = query_data
+#     return lines
 
 
 # TODO move to common
 def process_extracted_data(db_connection, row_data):
     for row_number, row_values in row_data.items():
         try:
-            record_exists = does_record_exist(db_connection, row_values)
+            record_exists = database.does_record_exist(db_connection, row_values)
             if common.valid_action(record_exists, row_values):
                 query, data = generate_db_query(row_values)
                 execute_db_query(db_connection, query, data, row_number, row_values)
@@ -70,58 +70,23 @@ def process_extracted_data(db_connection, row_data):
             raise e
 
 
-# TODO move to database with parameterised query or table name
-def does_record_exist(db, row_dict):
-    """
-    Checks to see if symptom group already exists in db with the id
-    """
-    record_exists = False
-    try:
-        with db.cursor(cursor_factory=psycopg2.extras.DictCursor) as cursor:
-            select_query = """select * from pathwaysdos.symptomgroups where id=%s"""
-            cursor.execute(select_query, (row_dict["id"],))
-            if cursor.rowcount != 0:
-                record_exists = True
-    except (Exception, psycopg2.Error) as e:
-        logger.log_for_error(
-            "Select symptom group by id failed - {0} => {1}".format(row_dict["id"], str(e)),
-        )
-        raise e
-    return record_exists
-
-
-def extract_query_data_from_csv(line):
+def extract_query_data_from_csv(lines):
     """
     Checks  maps data to db cols if correct
     """
-    csv_dict = {}
-    if common.check_csv_format(line, csv_column_count) and check_csv_values(line):
+    query_data = {}
+    for row_number, row_data in lines.items():
+
+        data_dict = {}
         try:
-            csv_dict["id"] = int(line[0])
-            csv_dict["name"] = line[1]
-            csv_dict["zcode"] = line[1].startswith("z2.0 - ")
-            csv_dict["action"] = line[2].upper()
+            data_dict["id"] = row_data["id"]
+            data_dict["name"] = row_data["description"]
+            data_dict["zcode"] = row_data["description"].startswith("z2.0 - ")
+            data_dict["action"] = row_data["action"].upper()
         except Exception as ex:
             logger.log_for_audit("CSV data invalid " + ex)
-    return csv_dict
-
-
-# TODO move to common maybe
-def check_csv_values(line):
-    """Returns false if either id or name are null or empty string"""
-    valid_values = True
-    try:
-        int(line[0])
-    except ValueError:
-        logger.log_for_audit("Id {} must be a integer".format(line[0]))
-        valid_values = False
-    if not str(line[0]):
-        logger.log_for_audit("Id {} can not be null or empty".format(line[0]))
-        valid_values = False
-    if not line[1]:
-        logger.log_for_audit("Name/Description {} can not be null or empty".format(line[1]))
-        valid_values = False
-    return valid_values
+        query_data[str(row_number)] = data_dict
+    return query_data
 
 
 # TODO move to util but call back to here for query content
@@ -173,7 +138,7 @@ def execute_db_query(db_connection, query, data, line, values):
     try:
         cursor.execute(query, data)
         db_connection.commit()
-        increment_summary_count(values)
+        common.increment_summary_count(summary_count_dict, values)
         logger.log_for_audit(
             "Action: {}, ID: {}, for symptomgroup {}".format(values["action"], values["id"], values["name"])
         )
@@ -183,22 +148,3 @@ def execute_db_query(db_connection, query, data, line, values):
         db_connection.rollback()
     finally:
         cursor.close()
-
-
-# TODO move to util if other jobs report counts
-def initialise_summary_count():
-    summary_count_dict[create_action] = 0
-    summary_count_dict[update_action] = 0
-    summary_count_dict[delete_action] = 0
-
-
-#  TODO move to util if other jobs report counts
-def increment_summary_count(values):
-    if values["action"] in [create_action, update_action, delete_action]:
-        summary_count_dict[values["action"]] = summary_count_dict[values["action"]] + 1
-    else:
-        logger.log_for_error(
-            "Can't increment count for action {0}. Valid actions are {1},{2},{3}".format(
-                values["action"], create_action, update_action, delete_action
-            )
-        )
