+++ application/utilities/common.py
@@ -1,10 +1,13 @@
-from utilities.database import DB
 from utilities.s3 import S3
 from utilities.logger import log_for_audit, log_for_error  # noqa
 from utilities.message import send_success_slack_message, send_failure_slack_message
+import csv
+
+create_action = "CREATE"
+update_action = "UPDATE"
+delete_action = "DELETE"
 
 
-# TODO rename as common
 def check_csv_format(csv_row, csv_column_count):
     """Checks length of csv data"""
     if len(csv_row) == csv_column_count:
@@ -41,17 +44,59 @@ def cleanup(db_connection, bucket, filename, event, start):
     return "Cleanup Successful"
 
 
-def connect_to_database(env, event, start):
-    db = DB()
-    log_for_audit("Setting DB connection details")
-    if not db.db_set_connection_details(env, event, start):
-        log_for_error("Error DB Parameter(s) not found in secrets store.")
-        send_failure_slack_message(event, start)
-        raise ValueError("DB Parameter(s) not found in secrets store")
-    return db.db_connect(event, start)
-
-
 def retrieve_file_from_bucket(bucket, filename, event, start):
     log_for_audit("Looking in {} for {} file".format(bucket, filename))
     s3_bucket = S3()
     return s3_bucket.get_object(bucket, filename, event, start)
+
+
+def check_csv_values(line):
+    """Returns false if either id or name are null or empty string"""
+    valid_values = True
+    try:
+        int(line[0])
+    except ValueError:
+        log_for_audit("Id {} must be a integer".format(line[0]))
+        valid_values = False
+    if not str(line[0]):
+        log_for_audit("Id {} can not be null or empty".format(line[0]))
+        valid_values = False
+    if not line[1]:
+        log_for_audit("Name/Description {} can not be null or empty".format(line[1]))
+        valid_values = False
+    return valid_values
+
+
+def initialise_summary_count():
+    summary_count_dict = {}
+    summary_count_dict[create_action] = 0
+    summary_count_dict[update_action] = 0
+    summary_count_dict[delete_action] = 0
+    return summary_count_dict
+
+
+def increment_summary_count(summary_count_dict, values):
+    if values["action"] in [create_action, update_action, delete_action]:
+        summary_count_dict[values["action"]] = summary_count_dict[values["action"]] + 1
+    else:
+        log_for_error(
+            "Can't increment count for action {0}. Valid actions are {1},{2},{3}".format(
+                values["action"], create_action, update_action, delete_action
+            )
+        )
+
+
+def process_file(csv_file, event, start, expected_col_count):
+    """returns dictionary of row data keyed on row number col1=id, col2=description, col3=action"""
+    lines = {}
+    count = 0
+    csv_reader = csv.reader(csv_file.split("\n"))
+    for line in csv_reader:
+        count += 1
+        if check_csv_format(line, expected_col_count) and check_csv_values(line):
+            lines[str(count)] = {"id": line[0], "description": line[1], "action": line[2]}
+        else:
+            log_for_error("Incorrect line format, should be {0} but is {1}".format(expected_col_count, len(line)))
+    if lines == {}:
+        send_failure_slack_message(event, start)
+    return lines
