+++ test/unit/hk/symptomgroups/test_hk_symptomgroups.py
@@ -4,7 +4,7 @@ import string
 import psycopg2
 
 from .. import handler
-from utilities import secrets
+from utilities import secrets,common
 
 file_path = "application.hk.symptomgroups.handler"
 filename = "test/sg.csv"
@@ -18,61 +18,76 @@ csv_sg_action = "REMOVE"
 
 def test_csv_line():
     """Test data extracted from valid csv"""
-    csv_row = [csv_sg_id, csv_sg_desc, csv_sg_action]
-    csv_dict = handler.extract_query_data_from_csv(csv_row)
-    assert len(csv_dict) == 4
-    assert csv_dict["id"] == csv_sg_id
-    assert csv_dict["name"] == str(csv_sg_desc)
-    assert csv_dict["action"] == str(csv_sg_action).upper()
-    assert csv_dict["zcode"] is False
+    # lines[str(count)] = {"id": line[0], "description": line[1], "action": line[2]}
+    csv_rows = {}
+    csv_rows["1"]={"id": csv_sg_id, "description": csv_sg_desc, "action": csv_sg_action}
+    csv_dict = handler.extract_query_data_from_csv(csv_rows)
+    assert len(csv_dict) == 1
+    assert len(csv_dict["1"]) == 4
+    assert csv_dict["1"]["id"] == csv_sg_id
+    assert csv_dict["1"]["name"] == str(csv_sg_desc)
+    assert csv_dict["1"]["action"] == str(csv_sg_action).upper()
+    assert csv_dict["1"]["zcode"] is False
 
 
 def test_csv_line_lc():
     """Test lower case action in csv is converted to u/c"""
+    csv_rows = {}
     csv_sg_action = "remove"
-    csv_row = [csv_sg_id, csv_sg_desc, csv_sg_action]
-    csv_dict = handler.extract_query_data_from_csv(csv_row)
-    assert len(csv_dict) == 4
-    assert csv_dict["id"] == csv_sg_id
-    assert csv_dict["name"] == str(csv_sg_desc)
-    assert csv_dict["action"] == str(csv_sg_action).upper()
-    assert csv_dict["zcode"] is False
+    # csv_row = [csv_sg_id, csv_sg_desc, csv_sg_action]
+    csv_rows["1"]={"id": csv_sg_id, "description": csv_sg_desc, "action": csv_sg_action}
+    csv_dict = handler.extract_query_data_from_csv(csv_rows)
+    assert len(csv_dict) == 1
+    assert len(csv_dict["1"]) == 4
+    assert csv_dict["1"]["id"] == csv_sg_id
+    assert csv_dict["1"]["name"] == str(csv_sg_desc)
+    assert csv_dict["1"]["action"] == str(csv_sg_action).upper()
+    assert csv_dict["1"]["zcode"] is False
+
+# no longer needed
+# def test_invalid_csv_line():
+#     """Test invalid csv results in no data extracted"""
+#     csv_rows = {}
+#     csv_row = [csv_sg_id, csv_sg_desc]
+#     csv_dict = handler.extract_query_data_from_csv(csv_row)
+#     assert len(csv_dict) == 0
 
 
-def test_invalid_csv_line():
-    """Test invalid csv results in no data extracted"""
-    csv_row = [csv_sg_id, csv_sg_desc]
-    csv_dict = handler.extract_query_data_from_csv(csv_row)
-    assert len(csv_dict) == 0
+# def test_no_sgid_csv_line():
+#     """Test no id in csv extract sets id to None"""
+#     csv_rows = {}
+#     # csv_row = ["", csv_sg_desc, csv_sg_action]
+#     csv_rows["1"]={"id": "", "description": csv_sg_desc, "action": csv_sg_action}
+#     csv_dict = handler.extract_query_data_from_csv(csv_row)
+#     assert len(csv_dict) == 0
 
 
-def test_no_sgid_csv_line():
-    """Test no id in csv extract sets id to None"""
-    csv_row = ["", csv_sg_desc, csv_sg_action]
-    csv_dict = handler.extract_query_data_from_csv(csv_row)
-    assert len(csv_dict) == 0
-
-
-def test_no_sgdesc_csv_line():
-    """Test if no name in csv extract sets name to None and zcode to False"""
-    csv_row = [csv_sg_id, "", csv_sg_action]
-    csv_dict = handler.extract_query_data_from_csv(csv_row)
-    assert len(csv_dict) == 0
+# def test_no_sgdesc_csv_line():
+#     """Test if no name in csv extract sets name to None and zcode to False"""
+#     csv_row = [csv_sg_id, "", csv_sg_action]
+#     csv_dict = handler.extract_query_data_from_csv(csv_row)
+#     assert len(csv_dict) == 0
 
 
 def test_zcode_sgdesc_csv_line():
     """Test extract correctly recognises zcodes"""
-    csv_row = [csv_sg_id, "z2.0 - test", csv_sg_action]
-    csv_dict = handler.extract_query_data_from_csv(csv_row)
-    assert len(csv_dict) == 4
-    assert csv_dict["zcode"] is True
+    csv_rows = {}
+    # csv_row = [csv_sg_id, "z2.0 - test", csv_sg_action]
+    csv_rows["1"]={"id": csv_sg_id, "description": "z2.0 - test", "action": csv_sg_action}
+    csv_dict = handler.extract_query_data_from_csv(csv_rows)
+    assert len(csv_dict) == 1
+    assert len(csv_dict["1"]) == 4
+    assert csv_dict["1"]["zcode"] is True
+
 
 
 def test_csv_line_exception():
     """Test exception handling by deliberately setting action to NOT a string"""
-    csv_row = [csv_sg_id, csv_sg_desc, 1]
+    csv_rows = {}
+    csv_rows["1"]={"id": csv_sg_id, "description": "z2.0 - test", "action": 1}
+    # csv_row = [csv_sg_id, csv_sg_desc, 1]
     with pytest.raises(Exception):
-        handler.extract_query_data_from_csv(csv_row)
+        handler.extract_query_data_from_csv(csv_rows)
 
 def test_generating_create_query():
     """Test creation of insert query and data arguments"""
@@ -137,70 +152,71 @@ def test_generating_query_with_invalid_action():
     with pytest.raises(psycopg2.DatabaseError):
         query, data = handler.generate_db_query(csv_dict)
 
-def test_extract_data_from_file_valid_length():
-    """Test one valid line of csv equals one row of extracted data"""
-    csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n"""
-    lines = handler.extract_data_from_file(csv_file) #, event, start)
-    assert len(lines) == 1
+# def test_process_file_valid_length():
+#     """Test one valid line of csv equals one row of extracted data"""
+#     csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n"""
+#     lines = handler.process_file(csv_file) #, event, start)
+#     assert len(lines) == 1
 
-def test_extract_data_from_file_valid_length_multiline():
-    """Test two valid lines of csv equals two rows of extracted data"""
-    csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n2001,"Automated update SymptomGroup","UPDATE"\n"""
-    lines = handler.extract_data_from_file(csv_file) #, event, start)
-    assert len(lines) == 2
+# def test_process_file_valid_length_multiline():
+#     """Test two valid lines of csv equals two rows of extracted data"""
+#     csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n2001,"Automated update SymptomGroup","UPDATE"\n"""
+#     lines = handler.process_file(csv_file) #, event, start)
+#     assert len(lines) == 2
 
-def test_extract_data_from_file_empty_second_line():
-    """Test data extraction ignores any empty line at end of file"""
-    csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n\n"""
-    lines = handler.extract_data_from_file(csv_file) #, event, start
-    assert len(lines) == 1
+# def test_process_file_empty_second_line():
+#     """Test data extraction ignores any empty line at end of file"""
+#     csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n\n"""
+#     lines = handler.process_file(csv_file) #, event, start
+#     assert len(lines) == 1
 
-def test_extract_data_from_file_empty_first_line():
-    """Test data extraction ignores any empty line at start of file"""
-    csv_file = """\n2001,"Automated insert SymptomGroup","CREATE"\n"""
-    lines = handler.extract_data_from_file(csv_file) #, event, start
-    assert len(lines) == 1
+# def test_process_file_empty_first_line():
+#     """Test data extraction ignores any empty line at start of file"""
+#     csv_file = """\n2001,"Automated insert SymptomGroup","CREATE"\n"""
+#     lines = handler.process_file(csv_file) #, event, start
+#     assert len(lines) == 1
 
-@patch(f"{file_path}.extract_query_data_from_csv", return_value={"id": "2001", "name": "Automated insert SymptomGroup","zcode":"None", "action": "CREATE"})
-def test_extract_data_from_file_three_lines_empty_second_line(mock_extract):
-    """Test data extraction ignores any empty line in middle of file"""
-    csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n\n\n2001,"Automated update SymptomGroup","UPDATE"\n"""
-    lines = handler.extract_data_from_file(csv_file) #, event, start
-    assert len(lines) == 2
-    assert mock_extract.call_count == 2
+# # @patch(f"{file_path}.extract_query_data_from_csv", return_value={"id": "2001", "name": "Automated insert SymptomGroup","zcode":"None", "action": "CREATE"})
+# def test_process_file_three_lines_empty_second_line():
+#     """Test data extraction ignores any empty line in middle of file"""
+#     csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n\n\n2001,"Automated update SymptomGroup","UPDATE"\n"""
+#     lines = handler.process_file(csv_file) #, event, start
+#     assert len(lines) == 2
+#     # assert mock_extract.call_count == 2
 
-def test_extract_data_from_file_incomplete_second_line():
-    """Test data extraction ignores a line itf it is incomplete"""
-    csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n2002,\n"""
-    lines = handler.extract_data_from_file(csv_file) #, event, start
-    assert len(lines)==1
-    assert lines["1"]["id"] == 2001
+# def test_process_file_incomplete_second_line():
+#     """Test data extraction ignores a line itf it is incomplete"""
+#     csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n2002,\n"""
+#     lines = handler.process_file(csv_file) #, event, start
+#     assert len(lines)==1
+#     assert lines["1"]["id"] == 2001
 
-def test_extract_data_from_file_incomplete_first_line():
-    """Test data extraction ingores first line if it is incomplete"""
-    csv_file = """2002,\n2001,"Automated insert SymptomGroup","CREATE"\n"""
-    lines = handler.extract_data_from_file(csv_file) #, event, start
-    assert len(lines)==1
-    assert lines["2"]["id"] == 2001
+# def test_process_file_incomplete_first_line():
+#     """Test data extraction ingores first line if it is incomplete"""
+#     csv_file = """2002,\n2001,"Automated insert SymptomGroup","CREATE"\n"""
+#     lines = handler.process_file(csv_file) #, event, start
+#     assert len(lines)==1
+#     assert lines["2"]["id"] == 2001
 
-@patch(f"{file_path}.extract_query_data_from_csv", return_value={"id": "2001", "name": "Automated insert SymptomGroup","zcode":"None", "action": "CREATE"})
-def test_extract_data_from_file_call_count(mock_extract):
-    """Test data extraction calls code to extract data from csv one per non empty line"""
-    csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n2001,"Automated update SymptomGroup","UPDATE"\n"""
-    lines = handler.extract_data_from_file(csv_file) #, event, start
-    assert len(lines) == 2
-    assert mock_extract.call_count == len(lines)
+# # @patch(f"{file_path}.extract_query_data_from_csv", return_value={"id": "2001", "name": "Automated insert SymptomGroup","zcode":"None", "action": "CREATE"})
+# def test_process_file_call_count():
+#     """Test data extraction calls code to extract data from csv one per non empty line"""
+#     csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n2001,"Automated update SymptomGroup","UPDATE"\n"""
+#     lines = handler.process_file(csv_file) #, event, start
+#     assert len(lines) == 2
+#     # assert mock_extract.call_count == len(lines)
 
-@patch(f"{file_path}.extract_query_data_from_csv", return_value={"id": "2001", "name": "Automated insert SymptomGroup","zcode":"None", "action": "CREATE"})
-def test_extract_data_from_file_call_count_inc_empty_line(mock_extract):
-    """Test data extraction calls code to extract data ignores empty line"""
-    csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n\n2001,"Automated update SymptomGroup","UPDATE"\n"""
-    lines = handler.extract_data_from_file(csv_file) #, event, start
-    assert len(lines) == 2
-    assert mock_extract.call_count == len(lines)
+# # @patch(f"{file_path}.extract_query_data_from_csv", return_value={"id": "2001", "name": "Automated insert SymptomGroup","zcode":"None", "action": "CREATE"})
+# def test_process_file_call_count_inc_empty_line():
+#     """Test data extraction calls code to extract data ignores empty line"""
+#     csv_file = """2001,"Automated insert SymptomGroup","CREATE"\n\n2001,"Automated update SymptomGroup","UPDATE"\n"""
+#     lines = handler.process_file(csv_file) #, event, start
+#     assert len(lines) == 2
+#     # assert mock_extract.call_count == len(lines)
 
+@patch(f"{file_path}.common.increment_summary_count")
 @patch("psycopg2.connect")
-def test_execute_db_query_success(mock_db_connect):
+def test_execute_db_query_success(mock_db_connect,mock_summary):
     """Test code to execute query successfully"""
     line = """2001,"New Symptom Group","CREATE"\n"""
     data = ("New Symptom Group", "None", 2001)
@@ -211,17 +227,11 @@ def test_execute_db_query_success(mock_db_connect):
     mock_db_connect.cursor.return_value.__enter__.return_value = "Success"
     query = """update pathwaysdos.symptomgroups set name = (%s), zcodeexists = (%s)
         where id = (%s);"""
-
-    handler.initialise_summary_count()
-    assert handler.summary_count_dict[handler.create_action] == 0
-    assert handler.summary_count_dict[handler.update_action] == 0
-    assert handler.summary_count_dict[handler.delete_action] == 0
     handler.execute_db_query(mock_db_connect, query, data, line, values)
     mock_db_connect.commit.assert_called_once()
+    mock_summary.assert_called_once()
     mock_db_connect.cursor().close.assert_called_once()
-    assert handler.summary_count_dict[handler.create_action] == 1
-    assert handler.summary_count_dict[handler.update_action] == 0
-    assert handler.summary_count_dict[handler.delete_action] == 0
+
 
 @patch("psycopg2.connect")
 def test_execute_db_query_failure(mock_db_connect):
@@ -240,7 +250,7 @@ def test_execute_db_query_failure(mock_db_connect):
 @patch(f"{file_path}.execute_db_query")
 @patch(f"{file_path}.generate_db_query",return_value=("query", "data"))
 @patch(f"{file_path}.common.valid_action", return_value=True)
-@patch(f"{file_path}.does_record_exist", return_value=True)
+@patch(f"{file_path}.database.does_record_exist", return_value=True)
 def test_process_extracted_data_single_record(mock_exist,mock_valid_action,mock_generate,mock_execute, mock_db_connect):
     """Test extracting data calls each downstream functions once for one record"""
     row_data = {}
@@ -260,7 +270,7 @@ def test_process_extracted_data_single_record(mock_exist,mock_valid_action,mock_
 @patch(f"{file_path}.execute_db_query")
 @patch(f"{file_path}.generate_db_query",return_value=("query", "data"))
 @patch(f"{file_path}.common.valid_action", return_value=True)
-@patch(f"{file_path}.does_record_exist", return_value=True)
+@patch(f"{file_path}.database.does_record_exist", return_value=True)
 def test_process_extracted_data_multiple_records(mock_exist,mock_valid_action,mock_generate,mock_execute, mock_db_connect):
     """Test extracting data calls each downstream functions once for each record"""
     row_data = {}
@@ -292,7 +302,7 @@ def test_process_extracted_data_error_check_exists_fails(mock_db_connect):
         handler.process_extracted_data(mock_db_connect, row_data)
 
 @patch("psycopg2.connect")
-@patch(f"{file_path}.does_record_exist", return_value=True)
+@patch(f"{file_path}.database.does_record_exist", return_value=True)
 def test_process_extracted_data_error_check_exists_passes(mock_exists,mock_db_connect):
     """Test error handling when extracting data and record exist check passes"""
     row_data = {}
@@ -307,41 +317,8 @@ def test_process_extracted_data_error_check_exists_passes(mock_exists,mock_db_co
         handler.process_extracted_data(mock_db_connect, row_data)
     assert mock_exists.call_count == 1
 
-@patch("psycopg2.connect")
-def test_record_exists_true(mock_db_connect):
-    """Test correct data passed to check record exists - returning true"""
-    csv_dict = {}
-    csv_dict["id"] = csv_sg_id
-    csv_dict["name"] = csv_sg_desc
-    csv_dict["action"] = "DELETE"
-    csv_dict["zcode"] = None
-    mock_db_connect.cursor.return_value.__enter__.return_value.rowcount = 1
-    assert handler.does_record_exist(mock_db_connect,csv_dict)
 
-@patch("psycopg2.connect")
-def test_does_record_exist_false(mock_db_connect):
-    """Test correct data passed to check record exists - returning false"""
-    csv_dict = {}
-    csv_dict["id"] = csv_sg_id
-    csv_dict["name"] = csv_sg_desc
-    csv_dict["action"] = "DELETE"
-    csv_dict["zcode"] = None
-    mock_db_connect.cursor.return_value.__enter__.return_value.rowcount = 0
-    assert not handler.does_record_exist(mock_db_connect,csv_dict)
-
-@patch("psycopg2.connect")
-def test_does_record_exist_exception(mock_db_connect):
-    """Test throwing of exception """
-    csv_dict = {}
-    csv_dict["id"] = csv_sg_id
-    csv_dict["name"] = csv_sg_desc
-    csv_dict["action"] = "DELETE"
-    csv_dict["zcode"] = None
-    mock_db_connect = ""
-    with pytest.raises(Exception):
-        handler.does_record_exist(mock_db_connect,csv_dict)
-
-@patch(f"{file_path}.common.connect_to_database", return_value="db_connection")
+@patch(f"{file_path}.database.connect_to_database", return_value="db_connection")
 @patch(f"{file_path}.message.send_failure_slack_message", return_value = None)
 @patch(f"{file_path}.message.send_start_message", return_value = None)
 @patch(f"{file_path}.common.retrieve_file_from_bucket", return_value = None)
@@ -354,9 +331,9 @@ def test_handler_exception(mock_db,mock_failure_message,mock_message_start,mock_
 @patch("psycopg2.connect")
 @patch(f"{file_path}.common.cleanup")
 @patch(f"{file_path}.execute_db_query")
-@patch(f"{file_path}.does_record_exist", return_value=True)
+@patch(f"{file_path}.database.does_record_exist", return_value=True)
 @patch(f"{file_path}.common.retrieve_file_from_bucket", return_value="""2001,"New Symptom Group","UPDATE"\n""")
-@patch(f"{file_path}.common.connect_to_database", return_value="db_connection")
+@patch(f"{file_path}.database.connect_to_database", return_value="db_connection")
 @patch(f"{file_path}.message.send_start_message")
 def test_handler_pass(mock_send_start_message,mock_db_details,mock_get_object,
 mock_does_record_exist,mock_execute_db_query,mock_cleanup,mock_db_connect):
@@ -371,9 +348,9 @@ mock_does_record_exist,mock_execute_db_query,mock_cleanup,mock_db_connect):
 
 @patch("psycopg2.connect")
 @patch(f"{file_path}.common.cleanup")
-@patch(f"{file_path}.does_record_exist", return_value=False)
+@patch(f"{file_path}.database.does_record_exist", return_value=False)
 @patch(f"{file_path}.common.retrieve_file_from_bucket", return_value="""2001,"New Symptom Group","UPDATE"\n""")
-@patch(f"{file_path}.common.connect_to_database", return_value="db_connection")
+@patch(f"{file_path}.database.connect_to_database", return_value="db_connection")
 @patch(f"{file_path}.message.send_start_message")
 def test_handler_fail(mock_send_start_message,mock_db_details,mock_get_object,
 mock_does_record_exist,mock_cleanup,mock_db_connect):
@@ -385,81 +362,62 @@ mock_does_record_exist,mock_cleanup,mock_db_connect):
     mock_cleanup.assert_called_once()
     mock_does_record_exist.assert_called_once()
 
-def test_initialise_summary_count():
-    """Test summary counts initialised correctly"""
-    handler.initialise_summary_count()
-    assert(len(handler.summary_count_dict) == 3)
-    assert handler.summary_count_dict[handler.create_action] == 0
-    assert handler.summary_count_dict[handler.update_action] == 0
-    assert handler.summary_count_dict[handler.delete_action] == 0
+# def test_initialise_summary_count():
+#     """Test summary counts initialised correctly"""
+#     handler.initialise_summary_count()
+#     assert(len(handler.summary_count_dict) == 3)
+#     assert handler.summary_count_dict[handler.create_action] == 0
+#     assert handler.summary_count_dict[handler.update_action] == 0
+#     assert handler.summary_count_dict[handler.delete_action] == 0
 
-def test_increment_summary_count_create():
-    """Test only create count incremented for create action """
-    handler.initialise_summary_count()
-    assert handler.summary_count_dict[handler.create_action] == 0
-    assert handler.summary_count_dict[handler.update_action] == 0
-    assert handler.summary_count_dict[handler.delete_action] == 0
-    values = {"action":"CREATE"}
-    handler.increment_summary_count(values)
-    assert handler.summary_count_dict[handler.create_action] == 1
-    assert handler.summary_count_dict[handler.update_action] == 0
-    assert handler.summary_count_dict[handler.delete_action] == 0
+# def test_increment_summary_count_create():
+#     """Test only create count incremented for create action """
+#     handler.initialise_summary_count()
+#     assert handler.summary_count_dict[handler.create_action] == 0
+#     assert handler.summary_count_dict[handler.update_action] == 0
+#     assert handler.summary_count_dict[handler.delete_action] == 0
+#     values = {"action":"CREATE"}
+#     handler.increment_summary_count(values)
+#     assert handler.summary_count_dict[handler.create_action] == 1
+#     assert handler.summary_count_dict[handler.update_action] == 0
+#     assert handler.summary_count_dict[handler.delete_action] == 0
 
-def test_increment_summary_count_update():
-    """Test only update count incremented for update action """
-    handler.initialise_summary_count()
-    assert handler.summary_count_dict[handler.create_action] == 0
-    assert handler.summary_count_dict[handler.update_action] == 0
-    assert handler.summary_count_dict[handler.delete_action] == 0
-    values = {"action":"UPDATE"}
-    handler.increment_summary_count(values)
-    assert handler.summary_count_dict[handler.create_action] == 0
-    assert handler.summary_count_dict[handler.update_action] == 1
-    assert handler.summary_count_dict[handler.delete_action] == 0
+# def test_increment_summary_count_update():
+#     """Test only update count incremented for update action """
+#     handler.initialise_summary_count()
+#     assert handler.summary_count_dict[handler.create_action] == 0
+#     assert handler.summary_count_dict[handler.update_action] == 0
+#     assert handler.summary_count_dict[handler.delete_action] == 0
+#     values = {"action":"UPDATE"}
+#     handler.increment_summary_count(values)
+#     assert handler.summary_count_dict[handler.create_action] == 0
+#     assert handler.summary_count_dict[handler.update_action] == 1
+#     assert handler.summary_count_dict[handler.delete_action] == 0
 
-def test_increment_summary_count_delete():
-    """Test only delete count incremented for delete action """
-    handler.initialise_summary_count()
-    assert handler.summary_count_dict[handler.create_action] == 0
-    assert handler.summary_count_dict[handler.update_action] == 0
-    assert handler.summary_count_dict[handler.delete_action] == 0
-    values = {"action":"DELETE"}
-    handler.increment_summary_count(values)
-    assert handler.summary_count_dict[handler.create_action] == 0
-    assert handler.summary_count_dict[handler.update_action] == 0
-    assert handler.summary_count_dict[handler.delete_action] == 1
+# def test_increment_summary_count_delete():
+#     """Test only delete count incremented for delete action """
+#     handler.initialise_summary_count()
+#     assert handler.summary_count_dict[handler.create_action] == 0
+#     assert handler.summary_count_dict[handler.update_action] == 0
+#     assert handler.summary_count_dict[handler.delete_action] == 0
+#     values = {"action":"DELETE"}
+#     handler.increment_summary_count(values)
+#     assert handler.summary_count_dict[handler.create_action] == 0
+#     assert handler.summary_count_dict[handler.update_action] == 0
+#     assert handler.summary_count_dict[handler.delete_action] == 1
 
-def test_increment_summary_count_nosuch():
-    """Test NO count incremented for invalid action """
-    handler.initialise_summary_count()
-    assert handler.summary_count_dict[handler.create_action] == 0
-    assert handler.summary_count_dict[handler.update_action] == 0
-    assert handler.summary_count_dict[handler.delete_action] == 0
-    values = {"action":"NOSUCH"}
-    handler.increment_summary_count(values)
-    assert handler.summary_count_dict[handler.create_action] == 0
-    assert handler.summary_count_dict[handler.update_action] == 0
-    assert handler.summary_count_dict[handler.delete_action] == 0
+# def test_increment_summary_count_nosuch():
+#     """Test NO count incremented for invalid action """
+#     handler.initialise_summary_count()
+#     assert handler.summary_count_dict[handler.create_action] == 0
+#     assert handler.summary_count_dict[handler.update_action] == 0
+#     assert handler.summary_count_dict[handler.delete_action] == 0
+#     values = {"action":"NOSUCH"}
+#     handler.increment_summary_count(values)
+#     assert handler.summary_count_dict[handler.create_action] == 0
+#     assert handler.summary_count_dict[handler.update_action] == 0
+#     assert handler.summary_count_dict[handler.delete_action] == 0
 
-def test_check_for_not_null_values():
-    """Checks key values not null"""
-    csv_line = [2001,"Not null Id","UPDATE"]
-    assert handler.check_csv_values(csv_line)
-
-def test_check_for_null_id():
-    """Checks if id is null/empty string"""
-    csv_line = ["","Null Id","UPDATE"]
-    assert not handler.check_csv_values(csv_line)
-
-def test_check_for_null_name():
-    """Checks is name/description is null/empty string"""
-    csv_line = [2001,"","UPDATE"]
-    assert not handler.check_csv_values(csv_line)
-
-def test_check_for_alpha_id():
-    """Checks is name/description is null/empty string"""
-    csv_line = ["abc","","UPDATE"]
-    assert not handler.check_csv_values(csv_line)
 
 def generate_event_payload():
     """Utility function to generate dummy event data"""
